did some refactoring things
