+++ app/src/main/java/rectangledbmi/com/pittsburghrealtimetracker/handlers/Constants.java
@@ -5,6 +5,8 @@ import android.os.Bundle;
 import java.text.SimpleDateFormat;
 import java.util.Locale;
 
+import rectangledbmi.com.pittsburghrealtimetracker.ui.activities.SelectTransit;
+
 /**
  * Global constants available to everyone
  *
@@ -29,7 +31,7 @@ public class Constants {
 
     /**
      * The default date format to parse... The timezone is set as EST in
-     * {@link rectangledbmi.com.pittsburghrealtimetracker.SelectTransit#onCreate(Bundle)}
+     * {@link SelectTransit#onCreate(Bundle)}
      * @since 46
      */
     public final static SimpleDateFormat DEFAULT_DATE_PARSE_FORMAT = new SimpleDateFormat(DATE_FORMAT_PARSE, Locale.US);
