+++ app/src/main/res/layout/activity_about.xml
@@ -3,5 +3,5 @@
     android:id="@+id/container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="rectangledbmi.com.pittsburghrealtimetracker.AboutActivity"
+    tools:context=".ui.activities.AboutActivity"
     tools:ignore="MergeRootFrame" />
