+++ app/src/main/res/layout/activity_select_transit.xml
@@ -5,7 +5,7 @@
     android:id="@+id/drawer_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SelectTransit">
+    tools:context=".ui.activities.SelectTransit">
 
     <!--<FrameLayout-->
         <!--android:layout_width="match_parent"-->
@@ -27,7 +27,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:id="@+id/map"
-        android:name="rectangledbmi.com.pittsburghrealtimetracker.BusMapFragment"
+        android:name="rectangledbmi.com.pittsburghrealtimetracker.ui.fragments.BusMapFragment"
         tools:layout="@layout/fragment_bus_map" />
 
     <!-- android:layout_gravity="start" tells DrawerLayout to treat
@@ -41,7 +41,7 @@
         android:layout_width="@dimen/navigation_drawer_width"
         android:layout_height="match_parent"
         android:layout_gravity="start"
-        android:name="rectangledbmi.com.pittsburghrealtimetracker.NavigationDrawerFragment"
+        android:name="rectangledbmi.com.pittsburghrealtimetracker.ui.fragments.NavigationDrawerFragment"
         tools:layout="@layout/navigation_drawer_layout" />
 
 </android.support.v4.widget.DrawerLayout>
