+++ app/src/main/res/layout/fragment_bus_map.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="rectangledbmi.com.pittsburghrealtimetracker.BusMapFragment">
+    tools:context=".ui.fragments.BusMapFragment">
 
     <com.google.android.gms.maps.MapView
         android:layout_width="match_parent"
