+++ app/src/main/res/layout/navigation_drawer_layout.xml
@@ -4,7 +4,7 @@
     android:orientation="vertical"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".NavigationDrawerFragment"
+    tools:context=".ui.fragments.NavigationDrawerFragment"
     android:theme="@style/AppTheme"
     android:id="@+id/new_navigation_drawer"
     >
