+++ app/src/main/res/menu/about.xml
@@ -1,6 +1,5 @@
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    tools:context="rectangledbmi.com.pittsburghrealtimetracker.AboutActivity" >
+<menu xmlns:tools="http://schemas.android.com/tools"
+    tools:context=".ui.activities.AboutActivity">
     <!--<item android:id="@+id/action_settings"-->
         <!--android:title="@string/action_settings"-->
         <!--android:orderInCategory="100"-->
