+++ app/src/main/res/menu/select_transit.xml
@@ -1,6 +1,6 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".SelectTransit"
+    tools:context=".ui.activities.SelectTransit"
     xmlns:appcompat="http://schemas.android.com/apk/res-auto">
     <item android:id="@+id/action_clear"
         android:title="@string/action_clear"
